This commit contains the following changes:  - Disabled the bounds checking in block residual functions since it causes problems when simulating optimization models where simulation trajectories goes outside of bounds, see #2197.  - Reverted r3818 and r3807 related to automatic scaling of equation systems since they caused problems when simulating some FMUs. This strategy needs to be analyzed further and tested on larger models, see #2200.
git-svn-id: https://svn.jmodelica.org@3856 7ae51c65-13d5-4921-9d80-6511ddd1419c

+++ trunk/JMI/src/jmi_block_residual.h
@@ -113,7 +113,6 @@ struct jmi_block_residual_t {
 	int n;                         /**< \brief The number of real unknowns in the equation system */
 	int n_nr;                         /**< \brief The number of non-real unknowns in the equation system */
 	jmi_real_t* x;                 /**< \brief Work vector for the real iteration variables */
-    
 	jmi_real_t* x_nr;                 /**< \brief Work vector for the non-real iteration variables */
 	jmi_real_t* dx;				   /**< \brief Work vector for the seed vector */
 	jmi_real_t* dv;					/**< \brief Work vector for (dF/dv)*dv */
@@ -138,9 +137,9 @@ struct jmi_block_residual_t {
     
     int init;			   /**< \brief A flag for initialization */
     
-    long int nb_calls;                    /**< \brief Nb of times the block has been solved */
-    long int nb_iters;                     /**< \breif Total nb if iterations of non-linear solver */
-    long int nb_jevals ;
+    int nb_calls;                    /**< \brief Nb of times the block has been solved */
+    int nb_iters;                     /**< \breif Total nb if iterations of non-linear solver */
+    int nb_jevals ;
     double time_spent;             /**< \brief Total time spent in non-linear solver */
 };
